# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Networking Platform API",
        "description": "This is the API documentation for the Social Networking Platform.",
        "contact": {
            "email": "support@socialnetworkingplatform.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/friend-request/send/{friendId}": {
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Send a friend request",
                "description": "Send a friend request to another user.",
                "operationId": "eb0e8815131bd8c664c76ee53a98a868",
                "parameters": [
                    {
                        "name": "friendId",
                        "in": "path",
                        "description": "ID of the user to send a friend request to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Friend request sent.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Friend request sent"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FriendshipResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error sending friend request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/friend-request/accept/{userId}": {
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Accept a friend request",
                "description": "Accept a pending friend request from another user.",
                "operationId": "dfa67cb082207c14b7b76a92a095cf8c",
                "parameters": [
                    {
                        "name": "friendId",
                        "in": "path",
                        "description": "ID of the user whose friend request is accepted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request accepted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Friend request accepted"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FriendshipResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error accepting friend request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/friend-request/reject/{userId}": {
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Reject a friend request",
                "description": "Reject a pending friend request from another user.",
                "operationId": "28845bd6d1d63262022dd8d7affe5a22",
                "parameters": [
                    {
                        "name": "friendId",
                        "in": "path",
                        "description": "ID of the user whose friend request is rejected",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request rejected.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Friend request rejected"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FriendshipResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error rejecting friend request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/friends": {
            "get": {
                "tags": [
                    "Friendships"
                ],
                "summary": "List all friends",
                "description": "Retrieve a list of all the friends of the authenticated user.",
                "operationId": "d56e0d858e672778539c219d050a6b11",
                "responses": {
                    "200": {
                        "description": "Friends list retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Friends list retrieved"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving friends list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/friends/pending": {
            "get": {
                "tags": [
                    "Friendships"
                ],
                "summary": "List all pending friend requests",
                "description": "Retrieve a list of all pending friend requests for the authenticated user.",
                "operationId": "467930aaf8bd4386f51df6e167b9099e",
                "responses": {
                    "200": {
                        "description": "Pending friend requests retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pending friend requests retrieved"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FriendshipResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving pending requests.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FriendshipResource": {
                "title": "Friendship Resource",
                "description": "Represents the friendship resource",
                "properties": {
                    "id": {
                        "description": "Friendship ID",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "The ID of the user who sent the request",
                        "type": "integer",
                        "example": 10
                    },
                    "friend_id": {
                        "description": "The ID of the user who received the request",
                        "type": "integer",
                        "example": 15
                    },
                    "status": {
                        "description": "The status of the friendship request",
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "description": "Timestamp when the friendship request was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-05T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the friendship request was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-05T12:45:56Z"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "User resource response",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJh..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}